backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-jwt-extended==4.6.0
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
python-dotenv==1.0.0
bcrypt==4.1.2
marshmallow==3.20.1
gunicorn==21.2.0
# psycopg2-binary==2.9.9  # Uncomment for PostgreSQL support

# Supabase
supabase==2.4.0
httpx[http2]>=0.24,<0.26
python-jose[cryptography]==3.3.0
